% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{basic_index}
\alias{basic_index}
\title{Index function for basic model}
\usage{
basic_index(info)
}
\arguments{
\item{info}{A \code{list} object containing \code{info} from generated ZamCovid model.}
}
\value{
A \code{list} with elements \code{run} and \code{state}
}
\description{
Index function for basic model
}
\examples{
basic_index(info)
}
