% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{basic_transform}
\alias{basic_transform}
\title{Make transform function for basic model parameters}
\usage{
basic_transform(baseline)
}
\arguments{
\item{baseline}{A \code{list} containing all baseline parameters.}
}
\value{
A \code{list} of \code{odin} parameters for the basic model.
}
\description{
Make transform function for basic model parameters
}
\examples{
basic_transform(baseline)
}
