% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccination.R
\name{vaccine_schedule_historic}
\alias{vaccine_schedule_historic}
\title{Create historic vaccine schedule}
\usage{
vaccine_schedule_historic(
  data = NULL,
  uptake = NULL,
  age_priority = NULL,
  pop_to_vaccinate = NULL,
  daily_doses_value = NULL,
  weekly_doses = FALSE,
  days_between_doses = NULL,
  start = NULL,
  dose_waste = 0,
  n_days = NULL
)
}
\arguments{
\item{data}{A data.frame with columns \code{date}, \code{age_band_min},
and numbered doses columns, e.g. if there are two doses
these should be \code{dose1} and \code{dose2}. Values of
\code{age_band_min} must be multiples of 5 from 0 to 75.}

\item{uptake}{A matrix of 16 rows, and number of columns equal to
number of doses. The (i,j)th entry gives the fractional uptake
of dose j for group i. Should be non-increasing across rows.}
}
\value{
A \link{vaccine_schedule} object
}
\description{
Create a hypothetical historic vaccine schedule.
}
