% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{ZamCovid_check_data}
\alias{ZamCovid_check_data}
\title{Check data for particle filter}
\usage{
ZamCovid_check_data(data)
}
\arguments{
\item{data}{A data.frame of data}
}
\value{
Invisibly, a data frame, identical to \code{data}
}
\description{
Check that data for the particle filter has required columns and apply
constraints for fitting to it.
}
