% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{ZamCovid_index}
\alias{ZamCovid_index}
\title{Index function for ZamCovid model}
\usage{
ZamCovid_index(info)
}
\arguments{
\item{info}{A \code{list} object containing \code{info} from generated ZamCovid model.}
}
\value{
A \code{list} with elements \code{run} and \code{state}
}
\description{
Index function for ZamCovid model
}
\examples{
basic_index(info)
}
