% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{ZamCovid_initial}
\alias{ZamCovid_initial}
\title{Initial conditions for the ZamCovid model}
\usage{
ZamCovid_initial(info, n_particles, pars)
}
\value{
A numeric vector of initial conditions
}
\description{
Create initial conditions for the ZamCovid model. This matches the
interface required for mcstate
}
\examples{
p <- ZamCovid_parameters(numeric_date("2020-02-07"))
mod <- ZamCovid$new(p, 0, 10)
ZamCovid_initial(mod$info(), 10, p, 4L)
}
