% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{zamcovid_new_pars}
\alias{zamcovid_new_pars}
\title{Recover fitted parameters model samples}
\usage{
zamcovid_new_pars(samples, priors)
}
\arguments{
\item{samples}{A \code{samples} object from a pMCMC model run.}

\item{priors}{\code{list} of priors used for model run.}
}
\value{
A \code{list} of new model priors and vcv matrix.
}
\description{
Recover fitted parameters model samples
}
