% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{ZamCovid_parameters}
\alias{ZamCovid_parameters}
\title{Define ZamCovid model parameters}
\usage{
ZamCovid_parameters(
  start_date,
  steps_per_day = 4L,
  beta_date = NULL,
  beta_value = NULL,
  population = NULL,
  contact_matrix = NULL,
  N_tot = NULL,
  n_groups = NULL,
  severity = NULL,
  progression = NULL,
  vaccination = NULL,
  initial_seed_pattern = 1,
  initial_seed_size = 10,
  I_A_transmission = 0.223,
  I_P_transmission = 1,
  I_C_1_transmission = 1,
  I_C_2_transmission = 0,
  hosp_transmission = 0,
  G_D_transmission = 0,
  rel_susceptibility = 1,
  rel_p_sympt = 1,
  rel_p_hosp_if_sympt = 1,
  rel_p_death = 1,
  rel_infectivity = 1,
  vaccine_progression_rate = NULL,
  vaccine_schedule = NULL,
  vaccine_index_dose2 = NULL,
  vaccine_index_booster = NULL,
  vaccine_catchup_fraction = 1,
  n_doses = 2L
)
}
\arguments{
\item{start_date}{A positive integer, greated than \code{model_start} for the model
end time.}

\item{steps_per_day}{A positive integer for the number of steps per day to
run the model.}

\item{population}{Either \code{NULL} or a \code{data.frame} with columns \code{age_group}
and \code{n} of same length (age groups) as \code{contact_matrix} supplied. If
\code{NULL} defaults to nationally representative values.}

\item{contact_matrix}{Either \code{NULL} or a \code{data.frame} of a symmetric
individual-to-individual contact rate matrix. If \code{NULL} defaults to
nationally representative values.}

\item{model_start}{Either 0 or a positive integer for the initial start time
for the model.}

\item{seed_pars}{Either \code{NULL} or a \code{list} with elements \code{seed_group},
\code{seed_number} and \code{seed_time} for the \code{age_group} in which initial
infections will be seeded, the number of infection to seed and the model
time in which they will be seeded, respectively. If \code{NULL} will
default to seeding 5 infection in \code{age_group} 5 at \code{t = 0}.}
}
\value{
Return a \code{list} of parameters to run the basic ZamCovid model.
}
\description{
Define ZamCovid model parameters
}
