% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{ZamCovid_parameters_observation}
\alias{ZamCovid_parameters_observation}
\title{ZamCovid sensitivity and specificity parameters}
\usage{
ZamCovid_parameters_observation(exp_noise = 1e+06)
}
\arguments{
\item{sero_specificity}{Specificity of the serology test assay}

\item{sero_sensitivity}{Sensitivity of the serology test assay}

\item{pcr_specificity}{Specificity of the PCR test}

\item{pcr_sensitivity}{Sensitivity of the PCR test}
}
\value{
A list of parameter values
}
\description{
ZamCovid diagnostic test performance parameters
}
