% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{ZamCovid_parameters_progression}
\alias{ZamCovid_parameters_progression}
\title{ZamCovid progression parameters}
\usage{
ZamCovid_parameters_progression(
  dt,
  gamma_E = NULL,
  gamma_A = NULL,
  gamma_P = NULL,
  gamma_C_1 = NULL,
  gamma_C_2 = NULL,
  gamma_H_sev = NULL,
  gamma_H_D = NULL,
  gamma_H_R = NULL,
  gamma_G_D = NULL,
  gamma_R = NULL,
  gamma_U = NULL,
  gamma_sero_pre = NULL,
  gamma_sero_pos = NULL,
  gamma_PCR_pre = NULL,
  gamma_PCR_pos = NULL
)
}
\arguments{
\item{dt}{The step size}

\item{gamma_E}{Time-varying mean duration in the E (exposed) compartment.}

\item{gamma_A}{Time-varying mean duration in the I_A (asymptomatic)
compartment.}

\item{gamma_P}{Time-varying mean duration in the I_P (pre-symptomatic)
compartment.}

\item{gamma_C_1}{Time-varying mean duration in the I_C_1 first symptomatic
compartment.}

\item{gamma_C_2}{Time-varying mean duration in the I_C_2 second symptomatic
compartment.}

\item{gamma_H_D}{Time-varying parameters for the Erlang rate parameter of
the duration in the H_D (hospitalisation leading to death) compartment.}

\item{gamma_H_R}{Time-varying parameters for the Erlang rate parameter of
the duration in the H_R (hospitalisation leading to recovery) compartment.}

\item{gamma_G_D}{Time-varying parameters for the Erlang rate parameter of
the duration in the G_D (death in the community) compartment.}

\item{gamma_R}{Time-varying mean duration in the R (recovered) compartment;
effectively duration of infection-induced immunity.}

\item{gamma_U}{Time-varying mean delay from hospital admission to diagnosis
for those not confirmed on admission.}

\item{gamma_sero_pre}{Time-varying parameters for the Erlang rate parameter
of the time from exposed to becoming seropositive.}

\item{gamma_sero_pos}{Time-varying parameters for the Erlang rate parameter
of the duration of seropositivity.}

\item{gamma_PCR_pre}{Time-varying parameters for the Erlang rate parameter
of the time from infection exposure to becoming PCR positive.}

\item{gamma_PCR_pos}{Time-varying parameters for the Erlang rate parameter
of the duration of PCR positivity.}
}
\value{
A list of parameter values
}
\description{
ZamCovid progression parameters.  The \code{k_} parameters are the
scaling parameters for Erlang distributed progressions, while
the \code{gamma_} parameters are either the gamma parameters of that
distribution or the mean of exponentially distributed progressions.
}
\section{Time-varying parameters}{

Every time varying parameter has the same format, which can be \code{NULL} (in
which case a default single value is used) or a list with \code{date} and
\code{value} for changes in the parameter. If \code{value} is scalar then
\code{date} can be \code{NULL} or missing. If \code{value} is a vector then \code{date}
must be a vector of numeric dates of the same length as \code{value}.
}

