% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{ZamCovid_parameters_sens_and_spec}
\alias{ZamCovid_parameters_sens_and_spec}
\title{ZamCovid sensitivity and specificity parameters}
\usage{
ZamCovid_parameters_sens_and_spec(
  sero_specificity = 0.999,
  sero_sensitivity = 0.927,
  PCR_specificity = 0.99,
  PCR_sensitivity = 0.99
)
}
\arguments{
\item{sero_specificity}{Specificity of the serology test assay}

\item{sero_sensitivity}{Sensitivity of the serology test assay}

\item{PCR_specificity}{Specificity of the PCR test}

\item{PCR_sensitivity}{Sensitivity of the PCR test}
}
\value{
A list of parameter values
}
\description{
ZamCovid diagnostic test performance parameters
}
