% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{ZamCovid_parameters_vaccination}
\alias{ZamCovid_parameters_vaccination}
\title{ZamCovid vaccination parameters}
\usage{
ZamCovid_parameters_vaccination(
  dt,
  N_tot,
  n_groups,
  n_doses,
  rel_susceptibility,
  rel_p_sympt,
  rel_p_hosp_if_sympt,
  rel_p_death,
  rel_infectivity,
  vaccine_progression_rate,
  vaccine_schedule,
  vaccine_index_dose2,
  vaccine_index_booster,
  vaccine_catchup_fraction = 1
)
}
\arguments{
\item{dt}{The step size}
}
\value{
A list of vaccination parameters
}
\description{
ZamCovid vaccination parameters
}
