% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{zamcovid_plot_traces}
\alias{zamcovid_plot_traces}
\title{Plot basci ZamCovid pMCMC traces.}
\usage{
zamcovid_plot_traces(samples)
}
\arguments{
\item{samples}{A \code{list} object containing the sample outputs of a basic
model run.}
}
\value{
A base R plot of the pMCMC chains for fitted parameters and
log-likelihood.
}
\description{
Plot basci ZamCovid pMCMC traces.
}
