% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZamCovid_Rt.R
\name{ZamCovid_Rt}
\alias{ZamCovid_Rt}
\title{Compute "Rt"}
\usage{
ZamCovid_Rt(
  time,
  S,
  p,
  type = NULL,
  eigen_method = "power_iteration",
  R = NULL,
  weight_Rt = FALSE
)
}
\arguments{
\item{time}{A vector of time steps that the model was run over}

\item{S}{A (n groups x n vaccine classes) x steps matrix of "S"
compartment counts}

\item{p}{A \code{\link[=ZamCovid_parameters]{ZamCovid_parameters()}} object}

\item{type}{A character vector of possible Rt types to
compute. Can be \code{eff_Rt_general} or \code{Rt_general}, for Rt accounting or
not accounting for immunity in the population, respectively.}

\item{eigen_method}{The eigenvalue method to use (passed to
\link[eigen1:eigen1]{eigen1::eigen1} as \code{method})}

\item{R}{A (n groups x n strains x n vaccine classes) x time steps matrix of
"R" compartment counts, required for multi-strain models.}

\item{weight_Rt}{If \code{TRUE} then computes the weighted average
of the Rt for all strains, otherwise all calculations are returned with
an additional dimension to index each strain.}
}
\value{
A list with elements \code{time}, \code{beta}, and any of the \code{type}
values specified above.
}
\description{
Compute "Rt" for a single simulated trajectory and parameter set.
}
