% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{zamcovid_transform}
\alias{zamcovid_transform}
\title{Make transform function for ZamCovid basic model parameters}
\usage{
zamcovid_transform(baseline)
}
\arguments{
\item{baseline}{A \code{list} containing all baseline parameters.}
}
\value{
A \code{list} of \code{odin} parameters for the basic model.
}
\description{
Make transform function for ZamCovid basic model parameters
}
