% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{zamcovid_parameters}
\alias{zamcovid_parameters}
\title{Define ZamCovid basic model parameters}
\usage{
zamcovid_parameters(
  model_end,
  steps_per_day = 1,
  model_start = 0,
  contact_matrix = NULL,
  population = NULL,
  seed_pars = NULL,
  par_list = NULL
)
}
\arguments{
\item{model_end}{A positive integer, greated than \code{model_start} for the model
end time.}

\item{steps_per_day}{A positive integer for the number of steps per day to
run the model.}

\item{model_start}{Either 0 or a positive integer for the initial start time
for the model.}

\item{contact_matrix}{Either \code{NULL} or a \code{data.frame} of a symmetric
individual-to-individual contact rate matrix. If \code{NULL} defaults to
nationally representative values.}

\item{population}{Either \code{NULL} or a \code{data.frame} with columns \code{age_group}
and \code{n} of same length (age groups) as \code{contact_matrix} supplied. If
\code{NULL} defaults to nationally representative values.}

\item{seed_pars}{Either \code{NULL} or a \code{list} with elements \code{seed_group},
\code{seed_number} and \code{seed_time} for the \code{age_group} in which initial
infections will be seeded, the number of infection to seed and the model
time in which they will be seeded, respectively. If \code{NULL} will
default to seeding 5 infection in \code{age_group} 5 at \code{t = 0}.}

\item{pars_list}{Either \code{NULL} or a \code{list} of all other model parameters to
fix. If \code{NULL} default values will be supplied.}
}
\value{
Return a \code{list} of parameters to run the basic ZamCovid model.
}
\description{
Define ZamCovid basic model parameters
}
