% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{zamcovid_parameters}
\alias{zamcovid_parameters}
\title{Title}
\usage{
zamcovid_parameters(
  steps_per_day,
  date_start,
  date_end,
  contact_matrix = NULL,
  population = NULL,
  seed_group = 5,
  seed_number = 5,
  par_list = NULL
)
}
\arguments{
\item{dt}{}
}
\description{
Title
}
